[-] FO : Access address

+++ controllers/front/AddressController.php
@@ -66,6 +66,21 @@ class AddressControllerCore extends FrontController
                 $this->should_redirect = true;
             }
         } elseif (($id_address = (int)Tools::getValue('id_address'))) {
+
+            $addressForm = $this->address_form->loadAddressById($id_address);
+
+            if ($addressForm->getAddress()->id === null) {
+                return Tools::redirect('index.php?controller=404');
+            }
+
+            if (!$this->context->customer->isLogged()) {
+                return $this->redirectWithNotifications('/index.php?controller=authentication');
+            }
+
+            if ($addressForm->getAddress()->id_customer != $this->context->customer->id) {
+                return Tools::redirect('index.php?controller=404');
+            }
+
             if (Tools::getValue('delete')) {
                 $ok = $this->makeAddressPersister()->delete(
                     new Address($id_address, $this->context->language->id),
@@ -78,7 +93,6 @@ class AddressControllerCore extends FrontController
                     $this->errors[] = $this->l('Could not delete address.');
                 }
             } else {
-                $this->address_form->loadAddressById($id_address);
                 $this->context->smarty->assign('editing', true);
             }
         }
